result of
egrep -liR "the class of the player is (a )?" Trap\ Quest.inform/Inform/Extensions/ | while read f; do
sed -i -r 's/the class of the player is (not )?(an? |the )/the player is \1\2/g' "$f"
sed -i -r 's/the class of the player is (not )? ([aeouiAEOUI])/the player is \1an \2/g' "$f"
sed -i -r 's/the class of the player is (not )?/the player is \1a /g' "$f"
done


Signed-off-by: picobyte <PiC.o0.o.BytE@gmail.com>